% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{classified_data}
\alias{classified_data}
\title{Classification Based on Four Classes}
\usage{
classified_data(data)
}
\description{
Classification Based on Four Classes
}
\details{
This function returns a dataset named "data" in which includes an additional column that classifies the COICOP code for each code based on Durable, Semi-Durable, Non-Durable and Service goods.
}
